First Prize Certificate
I wrote an algorithm and developed the model to study the fate of co-operative nature of the society. Based on these results, I wrote a paper “Study of the growth of a particular strategy in the pool of strategies” whose primary objective was to provide the quantitative analysis of representative strategies played by individuals in the society. This paper won the first prize in the Pune Vidyarthi Gruha’s COET Techcraft 2013. I further extended this idea to model the interactions of the community as a whole and drafted the paper “Mathematical model of ‘comparative development of social health of communities’ using Prisoners’ Dilemma”. I presented this paper in the inter-college National level paper presentation competition held by Mind Spark, College of Engineering, Pune (COEP). The judges liked this idea and my paper won first prize among the top 100 teams which presented their ideas.